Extract Dormant to the Cycles namespace
